feat：Move `FeedViewModel`、`FeedLoadingViewModel`、`FeedErrorViewModel` to separate file
