+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedLoadingViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedLoadingViewModel.swift
+//  EssentialFeed
+//
+//  Created by nicaho on 2024/11/11.
+//
+
+import Foundation
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
