+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedViewModel.swift
+//  EssentialFeed
+//
+//  Created by nicaho on 2024/11/11.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
